Refactor into Next.js API route
