+++ package.json
@@ -10,9 +10,7 @@
   },
   "dependencies": {
     "axios": "^1.7.2",
-    "body-parser": "^1.20.2",
     "dotenv": "^16.4.5",
-    "express": "^4.19.2",
     "next": "^15.3.4",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
