+++ pages/api/mcp.js
@@ -1,26 +1,41 @@
-[
-  {
-    "name": "spotifyNowPlaying",
-    "description": "Returns the currently playing track on Jenny's Spotify.",
-    "type": "streamable_http",
-    "url": "https://spotifynowplaying-olbjl5wcvq-uc.a.run.app"
-  },
-  {
-    "name": "weatherNow",
-    "description": "Returns the current weather in West Vancouver.",
-    "type": "streamable_http",
-    "url": "https://weathernow-olbjl5wcvq-uc.a.run.app"
-  },
-  {
-    "name": "trafficToStore",
-    "description": "Estimates traffic time to Walmart or Save-On-Foods.",
-    "type": "streamable_http",
-    "url": "https://traffictostore-olbjl5wcvq-uc.a.run.app"
-  },
-  {
-    "name": "blinkStatus",
-    "description": "Returns the latest motion event from your Blink camera.",
-    "type": "streamable_http",
-    "url": "https://blinkstatus-olbjl5wcvq-uc.a.run.app"
+import { getSpotifySong } from '../../utils/spotify';
+import { getWeather } from '../../utils/weather';
+import { getTraffic } from '../../utils/traffic';
+import { getBlinkStatus } from '../../utils/blink';
+
+export default async function handler(req, res) {
+  if (req.method !== 'POST') {
+    res.setHeader('Allow', ['POST']);
+    return res.status(405).json({ error: 'Method not allowed' });
   }
-]
+
+  const { tool, query, parameters } = req.body;
+  let result = '';
+  let data = {};
+
+  try {
+    switch (tool) {
+      case 'spotify':
+        result = await getSpotifySong(query);
+        break;
+      case 'weather':
+        result = await getWeather(parameters?.location || 'West Vancouver');
+        break;
+      case 'traffic':
+        result = await getTraffic(parameters?.destination || 'Walmart');
+        break;
+      case 'blink':
+        result = await getBlinkStatus();
+        break;
+      default:
+        result = 'Unknown tool. Try again, reina.';
+    }
+    return res.status(200).json({ tool, response: result, data });
+  } catch (e) {
+    return res.status(500).json({
+      tool,
+      response: 'Error processing your request.',
+      error: e.toString(),
+    });
+  }
+}
